Networklint

+++ src/lib/network/network.cc
@@ -36,22 +36,22 @@ Descent_Socket CNetwork::socket(int domain, int type, int protocol) {
 }
 
 int CNetwork::close(Descent_Socket socket) {
-   int retval = 0;
+  int retval = 0;
 #ifdef _WIN32
-   retval = closesocket(socket);
+  retval = closesocket(socket);
 #else
-   retval = ::close(socket);
+  retval = ::close(socket);
 #endif
-   return retval;
+  return retval;
 }
 
 void CNetwork::atoinet(const string &input, struct sockaddr_in *addr) {
 #ifdef _WIN32
-   long inAddress;
-   inAddress = inet_addr(input.c_str());
-   addr->sin_addr.s_addr = inAddress;
+  int32_t inAddress;
+  inAddress = inet_addr(input.c_str());
+  addr->sin_addr.s_addr = inAddress;
 #else
-   inet_aton(input.c_str(), (struct in_addr *)&addr->sin_addr.s_addr);
+  inet_aton(input.c_str(), (struct in_addr *)&addr->sin_addr.s_addr);
 #endif
 }
 
@@ -59,19 +59,24 @@ struct protoent *CNetwork::getprotobyname(const char *name) {
   return ::getprotobyname(name);
 }
 
-int CNetwork::setsockopt(Descent_Socket sockfd, int level, int optname, const void *optval, socklen_t optlen) {
+int CNetwork::setsockopt(Descent_Socket sockfd, int level, int optname,
+                         const void *optval, socklen_t optlen) {
   return ::setsockopt(sockfd, level, optname, optval, optlen);
 }
 
-ssize_t CNetwork::recvfrom(Descent_Socket sockfd, void *buf, size_t len, int flags, struct sockaddr *src_addr, socklen_t *addrlen) {
+ssize_t CNetwork::recvfrom(Descent_Socket sockfd, void *buf, size_t len,
+                           int flags, struct sockaddr *src_addr,
+                           socklen_t *addrlen) {
   return ::recvfrom(sockfd, buf, len, flags, src_addr, addrlen);
 }
 
-int CNetwork::bind(Descent_Socket sockfd, const struct sockaddr *addr, socklen_t addrlen) {
+int CNetwork::bind(Descent_Socket sockfd, const struct sockaddr *addr,
+                   socklen_t addrlen) {
   return ::bind(sockfd, addr, addrlen);
 }
 
-int CNetwork::select(int nfds, fd_set *readfds, fd_set *writefds, fd_set *exceptfds, struct timeval *timeout) {
+int CNetwork::select(int nfds, fd_set *readfds, fd_set *writefds,
+                     fd_set *exceptfds, struct timeval *timeout) {
   return ::select(nfds, readfds, writefds, exceptfds, timeout);
 }
